+++ services/tracking/tracking_server/src/main/java/org/changppo/tracking/repository/TrackingRepository.java
@@ -1,10 +1,8 @@
 package org.changppo.tracking.repository;
 
-import org.changppo.tracking.domain.Tracking;
+import org.changppo.tracking.domain.mongodb.Tracking;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
-
-import java.time.LocalDateTime;
 
 @Repository
 public interface TrackingRepository extends MongoRepository<Tracking, String> {
