+++ services/tracking/tracking_server/src/test/java/org/changppo/tracking/api/TrackingControllerTest.java
@@ -1,10 +1,8 @@
 package org.changppo.tracking.api;
 
-import de.flapdoodle.reverse.transitions.Start;
 import org.changppo.tracking.api.request.StartTrackingRequest;
 import org.changppo.tracking.api.response.TrackingResponse;
 import org.changppo.tracking.base.WithCustomMockUser;
-import org.changppo.tracking.domain.Scope;
 import org.junit.jupiter.api.Nested;
 import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -23,7 +21,6 @@ import org.springframework.data.geo.Point;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import static org.mockito.ArgumentMatchers.any;
